Change on the API REST to get the quotes.

+++ demo-api/src/main/java/com/lil/springperformance/api/repository/DeviceRepository.java
@@ -15,7 +15,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface DeviceRepository extends JpaRepository<Device, Long> {
+public interface DeviceRepository extends JpaRepository<Device, Integer> {
 
 
 }
